Deploy robots within the runner
Note that the runner spec passes with no changes, since the first robot
in the example ends up in the same place it started.

For that reason, this commit adds the second example in.  The output
doesn't match the spec here (since we don't yet handle lost robots), but
a manual check shows that this is correct.