Update position before direction
The position update depends on the direction, which means that, if we
update the direction first, then the position update could incorrectly
use the new direction.

In practice this doesn't matter, since all of the current instructions
change only the position or the direction, never both.  However, if
more instructions are added later, this could matter.