Handle blank lines in input
The spec doesn't say that there will be blank lines between the robots,
but the example shown has them.

In the spirit of being liberal in what we accept but conservative with
what we emit, this commit ignores any blank lines in the input.  That
means that our program works whether we have blank lines or not.