Lazily read lines from input
Previously, the runner would attempt to load the entire input into
memory before processing any of it.  For very large inputs, this could
lead to very high memory usage.  This is shown by running the following,
which generates an arbitrarily long input.  The memory used by the
process grows linearly and eventually runs out.

{ echo '5 3' & yes '1 1 E\nRFRFRFRF'; } | bin/mars

This commit changes the runner to lazily read lines from the input,
processing each line as it is read.  With this change, the memory usage
remains bounded.

Since the co-ordinates and instruction strings are bounded (at 50 and
100 respectively), this means that any input can be handled in bounded
memory.